feat: Private Figma files
This commit adds a feature to embed private Figma files as images or
PDFs.

Though that works, you should use the feature carefully because it
exposes your token to the internet. If you use PAT, someone could steal
it and peek your other private files. If you use OAuth2, someone could
use your client_id and client_secret so they can obtain a Figma API
token with your credentials. IMO the latter is not so dangerous, but if
you set "*" to the redirect uri ("Callback"), anyone have access to the
Storybook can pretend as your app.