feat: Drop PDF support
We used `react-pdf`, which internally uses `pdf-js`. The problem is,
the worker-loader triggered by `react-pdf` causes errors such as
https://github.com/pocka/storybook-addon-designs/issues/72 .
Of course, that error is by user misconfiguration but it's hard to
manage such kind of errors just for the minor feature.

In addition to that, I don't want to ship with a lot of dependencies
just for preview-ing. Storybook bundle tends to be big and we addon
authors should try to minimize bundle size as much as possible.

If you want the PDF preview support, please use storybook-addon-designs@v5.

BREAKING CHANGE: parameters with `pdf` type no longer works