fix: Disable offscreen option for Figspec designs
https://github.com/pocka/storybook-addon-designs/issues/122

When a story with `type: "experimental-figspec"` places on 2nd or
later tab, the preview does not render Figma design correctly.
This is due to size calculation not updating on `display` changes.

However, the fundamentaly fix for it requires ResizeObserver,
which is still relatively new API. I would like to avoid these "young"
APIs in libraries because of documentation and/or additional dependency
management.

The Figspec type viewer can't have an instant tab switching while force
re-render is on, although I don't think it's a deal breaker.