perf: Do not destroy rendered contents on tab changes
Almost all of embed-type this addon supports take a few seconds to load.
The problem is, when a user moves to another addon tab then returns to
a "Design" tab, the tab contents restarts the whole rendering process
including network access. In order to eliminate this unnecessary
loading, I changed the panel rendering so it remains on DOM once after a
user enters the "Design" tab on the story.

https://github.com/pocka/storybook-addon-designs/issues/153